feat: add Projects app
